requirements.txt
mcp
httpx
python-dotenv
schedule
python-telegram-bot
flask
